Implemented raw e2e functionality. Removed fe bundle dist folder from repo.

+++ shaleapps-interview-be/todo.model.go
@@ -2,7 +2,7 @@ package main
 
 // Todo is the model for a to-do item
 type Todo struct {
-	ID          int    `json:"Id"`
+	ID          int64  `json:"Id"`
 	Description string `json:"Description"`
 	IsComplete  bool   `json:"IsComplete"`
 }
